Test that Graphs and Operations are gc'd despite cyclic references.
I ran into this working on a patch where I added a __del__ method to
Operation, which caused garbage collection to be disabled for cyclic
references. This triggered a failure in PyOpTest.testCleanup but was
very hard to debug. This test should more clearly guard against making
this mistake.
Change: 154365215