[XLA] Make HloModule always have a config.
This removes the circular dependency when creating a HloModule, then a
HloModuleConfig with the help of the module's entry computation, then assigning
the config back to the module. Now we have to pass a config when creating a
module, or a default config gets created.

This allows removing quite a bit of boilerplate code.

PiperOrigin-RevId: 157059949