[XLA] Don't add constraints for computations we're not currently looking at.
TuplePointsToAnalysis is computed globally per module, so we add all
unconstrained buffers in that module, even if it's outside of the computation
we're currently running on. Then we proceed to propagate default layouts to all
those buffers and then throw the constraints away because they don't affect any
instruction in the current computation.

PiperOrigin-RevId: 157635564