Add make_ndarray, tensor_proto, and MetaGraphDef to tf api.
Since TensorProtos are part of the TensorFlow API, it makes sense
to also include the methods that generate and parse them.

Similarly, we write out MetaGraphDef protos in the summary writer,
so we should provide the proto as well.

This is part of an ongoing effort to have TensorBoard only consume
TensorFlow methods through the public api.

PiperOrigin-RevId: 157657564